# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: JJK RPG Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /players/me:
    get:
      operationId: getMyPlayer
      tags: [player]
      summary: Get current player full data
      responses:
        "200":
          description: Player data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "401":
          description: Not authenticated
...
  /players/me/profile:
    patch:
      operationId: updateMyProfile
      tags: [player]
      summary: Update player profile (avatar, banner, title, badge, frame, theme)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /players/{userId}:
    get:
      operationId: getPlayer
      tags: [player]
      summary: Get public player profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicPlayer"
        "404":
          description: Player not found
  /players/me/character:
    post:
      operationId: equipCharacter
      tags: [player]
      summary: Equip a character
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipCharacterInput"
      responses:
        "200":
          description: Character equipped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"

  # ── CHARACTERS ────────────────────────────────────
...
  /leaderboards/level:
    get:
      operationId: getLevelLeaderboard
      tags: [leaderboards]
      summary: Top players by level
      parameters:
        - name: limit
          in: query
          schema:
            type: number
            default: 20
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"
  /leaderboards/wealth:
    get:
      operationId: getWealthLeaderboard
      tags: [leaderboards]
      summary: Richest players by Yen
      parameters:
        - name: limit
          in: query
          schema:
            type: number
            default: 20
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"
  /leaderboards/pvp:
    get:
      operationId: getPvpLeaderboard
      tags: [leaderboards]
      summary: Top players by PvP wins
      parameters:
        - name: limit
          in: query
          schema:
            type: number
            default: 20
      responses:
        "200":
          description: Leaderboard
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LeaderboardEntry"

  # ── CASINO ────────────────────────────────────────
...
  /marketplace/listings:
    get:
      operationId: listMarketplace
      tags: [marketplace]
      summary: Get all active marketplace listings
      parameters:
        - name: type
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: number
            default: 50
      responses:
        "200":
          description: Listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MarketplaceListing"

    post:
      operationId: createListing
      tags: [marketplace]
      summary: Create a marketplace listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListingInput"
      responses:
        "201":
          description: Listing created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketplaceListing"
  /marketplace/listings/{id}:
    delete:
      operationId: deleteListing
      tags: [marketplace]
      summary: Remove a marketplace listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted
  /marketplace/listings/{id}/buy:
    post:
      operationId: buyListing
      tags: [marketplace]
      summary: Buy a marketplace listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Purchased
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseResult"

  # ── NOTIFICATIONS ─────────────────────────────────
...
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  schemas:
...
    Player:
      type: object
      required: [user_id, username, display_name, level, xp, xp_needed, rank, yen, hp, max_hp, cursed_energy, max_cursed_energy, attack, defense, speed, wins, losses, win_rate]
      properties:
        user_id:
          type: number
        username:
          type: ["string", "null"]
        display_name:
          type: string
        character_id:
          type: ["integer", "null"]
        level:
          type: number
        xp:
          type: number
        xp_needed:
          type: number
        rank:
          type: string
        yen:
          type: number
        hp:
          type: number
        max_hp:
          type: number
        cursed_energy:
          type: number
        max_cursed_energy:
          type: number
        attack:
          type: number
        defense:
          type: number
        speed:
          type: number
        wins:
          type: number
        losses:
          type: number
        win_rate:
          type: number
        techniques:
          type: array
          items:
            type: string
        inventory:
          type: array
          items:
            type: number
        last_daily:
          type: ["string", "null"]
        created_at:
          type: ["string", "null"]
        # Extended dashboard profile fields
        avatar_url:
          type: ["string", "null"]
        banner_url:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        equipped_title:
          type: ["string", "null"]
        equipped_badge:
          type: ["string", "null"]
        equipped_frame:
          type: ["string", "null"]
        theme:
          type: ["string", "null"]
        is_banned:
          type: boolean
        is_admin:
          type: boolean
        elixirs:
          type: number
        domain_name:
          type: ["string", "null"]
        domain_power:
          type: ["integer", "null"]
        domain_equipped:
          type: boolean
    PublicPlayer:
      type: object
      required: [user_id, display_name, level, rank, wins, losses, win_rate]
      properties:
        user_id:
          type: number
        display_name:
          type: string
        username:
          type: ["string", "null"]
        character_id:
          type: ["integer", "null"]
        level:
          type: number
        rank:
          type: string
        wins:
          type: number
        losses:
          type: number
        win_rate:
          type: number
        avatar_url:
          type: ["string", "null"]
        equipped_title:
          type: ["string", "null"]
    ProfileUpdate:
      type: object
      properties:
        avatar_url:
          type: ["string", "null"]
        banner_url:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        equipped_title:
          type: ["string", "null"]
        equipped_badge:
          type: ["string", "null"]
        equipped_frame:
          type: ["string", "null"]
        theme:
          type: ["string", "null"]
...
    EquipCharacterInput:
      type: object
      required: [character_id]
      properties:
        character_id:
          type: number

    # ── CHARACTERS ────────────────────────────────────
...
    ShopItem:
      type: object
      required: [id, name, description, price, type]
      properties:
        id:
          type: number
        name:
          type: string
        description:
          type: string
        price:
          type: number
        type:
          type: string
        effect:
          type: ["string", "null"]
        image_url:
          type: ["string", "null"]
        use_description:
          type: ["string", "null"]
...
    PurchaseResult:
      type: object
      required: [success, message, new_yen]
      properties:
        success:
          type: boolean
        message:
          type: string
        new_yen:
          type: number
        item:
          $ref: "#/components/schemas/ShopItem"

    # ── INVENTORY ─────────────────────────────────────
...
    LeaderboardEntry:
      type: object
      required: [rank_position, user_id, display_name, level, yen, wins, win_rate]
      properties:
        rank_position:
          type: number
        user_id:
          type: number
        display_name:
          type: string
        username:
          type: ["string", "null"]
        character_id:
          type: ["integer", "null"]
        level:
          type: number
        rank:
          type: string
        yen:
          type: number
        wins:
          type: number
        losses:
          type: number
        win_rate:
          type: number
        avatar_url:
          type: ["string", "null"]

    # ── CASINO ────────────────────────────────────────
...
    ListingInput:
      type: object
      required: [item_id, price]
      properties:
        item_id:
          type: number
        price:
          type: number
        quantity:
          type: number
    MarketplaceListing:
      type: object
      required: [id, seller_id, seller_name, item_id, price, status, created_at]
      properties:
        id:
          type: number
        seller_id:
          type: number
        seller_name:
          type: string
        item_id:
          type: number
        item_name:
          type: string
        item_type:
          type: string
        price:
          type: number
        quantity:
          type: number
        status:
          type: string
        created_at:
          type: string

    # ── NOTIFICATIONS ─────────────────────────────────
    Notification:
      type: object
      required: [id, user_id, type, title, message, read, created_at]
      properties:
        id:
          type: number
        user_id:
          type: number
        type:
          type: string
        title:
          type: string
        message:
          type: string
        read:
          type: boolean
        created_at:
          type: string

    # ── TRANSACTIONS ──────────────────────────────────
...